Metadata manager
This change introduces a class responsible for fetching fresh metadata
and exposing the most recent version of it. Apart from direct calls
to the refresh method, metadata manager also attempts to fetch new metadata
in the background every time the current one expires. `_expiration_time`,
set in the constructor, indicates the maximum age of metadata.