Add kafka4seastar's buildsystem based on CMake
This change introduces the CMakeLists.txt file containing the build rules
for the library. This project is dependent on Seastar and attempts to find
its installation on the machine. This path can be set manually by setting
the following cmake cached variables:

	CMAKE_PATH_PREFIX to /path/to/seastar/build/release

	CMAKE_MODULE_PREFIX to /path/to/seastar/cmake

For building unit tests, this CMakeLists uses the file provided in tests/unit.